refactor widget structure
