+++ src/actions/index.js
@@ -1,27 +1,18 @@
-import { receiveWeekNumber, fetchWeekNumber } from './weekNumber';
-
 import {
-  // onReceive,
+  refresh as refreshWeekNumber,
+} from '../widgets/WeekNumber/actions/weekNumber';
+import {
   refresh as refreshNewRelicErrors,
 } from '../widgets/NewRelicErrors/actions/newRelicErrors';
-
 import {
-  requestBugsHistory,
-  receiveBugsHistory,
-  fetchBugsHistory,
-} from './bugsHistory';
+  refresh as refreshBugsHistory,
+} from '../widgets/BugsHistory/actions/bugsHistory';
 
-export {
-  receiveWeekNumber,
-  fetchWeekNumber,
-  requestBugsHistory,
-  receiveBugsHistory,
-  fetchBugsHistory,
-};
-
-export const refreshAll = () => dispatch => {
-  // eslint-disable-line import/prefer-default-export
-  dispatch(fetchWeekNumber());
+// TODO: Possibly refactor using Observers pattern. This way we won`t have to add widgets here manually.
+const refreshAll = () => dispatch => {
+  dispatch(refreshWeekNumber());
   dispatch(refreshNewRelicErrors());
-  dispatch(fetchBugsHistory());
+  dispatch(refreshBugsHistory());
 };
+
+export { refreshAll }; // eslint-disable-line import/prefer-default-export
