+++ src/components/WidgetList/WidgetList.test.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { shallow } from 'enzyme';
 import WidgetList from './WidgetList';
-import WeekNumber from '../WeekNumber/WeekNumber';
+import WeekNumber from '../../widgets/WeekNumber/components/WeekNumber';
 import NewRelicErrors
   from '../../widgets/NewRelicErrors/components/NewRelicErrors';
-import BugsHistory from '../BugsHistory/BugsHistory';
+import BugsHistory from '../../widgets/BugsHistory/components/BugsHistory';
 
 describe('WidgetList component', () => {
   const widget = shallow(<WidgetList />);
