--- src/reducers/bugsHistory/bugsHistory.js
@@ -1,13 +1,0 @@
-import { handleActions } from 'redux-actions';
-import { receiveBugsHistory } from '../../actions';
-
-const bugsHistory = handleActions(
-  {
-    [receiveBugsHistory]: (state, { payload }) => ({ history: payload }),
-  },
-  {
-    history: [],
-  },
-);
-
-export default bugsHistory;
