+++ src/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
-import weekNumber from './weekNumber/weekNumber';
+import weekNumber from '../widgets/WeekNumber/reducers/weekNumber';
 import newRelicErrors from '../widgets/NewRelicErrors/reducers/newRelicErrors';
-import bugsHistory from './bugsHistory/bugsHistory';
+import bugsHistory from '../widgets/BugsHistory/reducers/bugsHistory';
 
 const appReducers = combineReducers({
   weekNumber,
