+++ src/widgets/BugsHistory/actions/bugsHistory.test.js
