+++ src/widgets/BugsHistory/components/BugsHistory.scss
