+++ src/widgets/BugsHistory/components/BugsHistory.test.js
