+++ src/widgets/BugsHistory/reducers/bugsHistory.js
@@ -1,0 +1,13 @@
+import { handleActions } from 'redux-actions';
+import { onReceive } from '../actions/bugsHistory';
+
+const reducer = handleActions(
+  {
+    [onReceive]: (state, { payload }) => ({ history: payload }),
+  },
+  {
+    history: [],
+  },
+);
+
+export default reducer;
