+++ src/widgets/BugsHistory/reducers/bugsHistory.test.js
