+++ src/widgets/NewRelicErrors/actions/newRelicErrors.js
@@ -2,7 +2,7 @@ import 'whatwg-fetch';
 import { createAction } from 'redux-actions';
 import fetchApi from '../../../lib/fetchApi';
 
-export const onReceive = createAction('NEWRELICERRORS_RECEIVE', json => ({
+export const onReceive = createAction('RECEIVE_NEWRELICERRORS', json => ({
   previous: json.data.previous,
   current: json.data.current,
 }));
