+++ src/widgets/WeekNumber/reducers/weekNumber.test.js
@@ -1,9 +1,9 @@
-import weekNumber from './weekNumber';
+import reducer from './weekNumber';
 
 describe('weekNumber reducer', () => {
   it('updates the state with the data from a RECEIVE_WEEKNUMBER action', () => {
     expect(
-      weekNumber(
+      reducer(
         {},
         {
           type: 'RECEIVE_WEEKNUMBER',
@@ -17,12 +17,12 @@ describe('weekNumber reducer', () => {
 
   it('does not handle other actions', () => {
     expect(
-      weekNumber(
+      reducer(
         {
           week: 0,
         },
         {
-          type: 'RECEIVE_BUGSDIFF',
+          type: 'OTHER_EVENT',
           payload: 9,
         },
       ),
